+++ src/services/postToLens.ts
@@ -7,15 +7,11 @@ import { Source } from '@/constants/enum.js';
 import { SITE_URL } from '@/constants/index.js';
 import { readChars } from '@/helpers/chars.js';
 import { createDummyPost } from '@/helpers/createDummyPost.js';
-import { enqueueErrorMessage } from '@/helpers/enqueueMessage.js';
-import { getPollFrameUrl } from '@/helpers/getPollFrameUrl.js';
 import { getUserLocale } from '@/helpers/getUserLocale.js';
-import { resolveLensOperationName, resolveLensQuery } from '@/helpers/resolveLensQuery.js';
 import { createIPFSMediaObject, resolveImageUrl, resolveVideoUrl } from '@/helpers/resolveMediaObjectUrl.js';
 import { resolveSourceName } from '@/helpers/resolveSourceName.js';
 import { LensPollProvider } from '@/providers/lens/Poll.js';
 import { LensSocialMediaProvider } from '@/providers/lens/SocialMedia.js';
-import type { Poll } from '@/providers/types/Poll.js';
 import { createPostTo } from '@/services/createPostTo.js';
 import { uploadToArweave } from '@/services/uploadToArweave.js';
 import { uploadFileToIPFS } from '@/services/uploadToIPFS.js';
@@ -50,7 +46,7 @@ interface Attachments {
     }>;
 }
 
-function createPayloadAttachments(images: MediaObject[], video: MediaObject | null): Attachments | undefined {
+export function createPayloadAttachments(images: MediaObject[], video: MediaObject | null): Attachments | undefined {
     if (
         images.some((image) => !resolveImageUrl(Source.Lens, image)) ||
         (video && !resolveVideoUrl(Source.Lens, video))
@@ -91,11 +87,6 @@ function createPayloadAttachments(images: MediaObject[], video: MediaObject | nu
                     }),
           }
         : undefined;
-}
-
-function createPayloadSharingLink(polls: Poll[] | undefined): string | undefined {
-    if (!polls?.length) return;
-    return getPollFrameUrl(polls[0].id, Source.Lens);
 }
 
 export function createPostMetadata(baseMetadata: BaseMetadata, attachments?: Attachments, sharingLink?: string) {
@@ -323,86 +314,4 @@ export async function postToLens(type: ComposeType, compositePost: CompositePost
     });
 
     return postTo(type, compositePost);
-}
-export interface LensSchedulePayload {
-    operationName: 'PostOnMomoka' | 'QuoteOnMomoka' | 'CommentOnMomoka';
-    variables: {
-        request: {
-            contentURI: `ar://${string}`;
-            quoteOn?: string;
-            commentOn?: string;
-        };
-    };
-    query: string;
-}
-
-export async function createLensSchedulePostPayload(
-    type: ComposeType,
-    compositePost: CompositePost,
-    isThread = false,
-): Promise<LensSchedulePayload> {
-    const { images, video, poll, chars, parentPost, postId } = compositePost;
-
-    const lensParentPost = parentPost.Lens;
-    const sourceName = resolveSourceName(Source.Lens);
-
-    const imageResults = await Promise.all(
-        images.map(async (media) => {
-            if (resolveImageUrl(Source.Lens, media)) return media;
-            return createIPFSMediaObject(await uploadFileToIPFS(media.file), media);
-        }),
-    );
-
-    const videoResult = video?.file ? createIPFSMediaObject(await uploadFileToIPFS(video.file), video) : null;
-
-    const pollResult = !poll ? [] : [await LensPollProvider.createPoll(poll, readChars(chars, 'both', Source.Lens))];
-
-    const { currentProfile } = useLensStateStore.getState();
-    if (!currentProfile?.profileId) throw new Error(t`Login required to schedule post on ${sourceName}`);
-    if (!currentProfile.signless) {
-        const message = t`Please enable Momoka to support sending posts on Lens.`;
-        enqueueErrorMessage(message);
-        throw new Error(message);
-    }
-
-    const title = `Post by #${currentProfile.handle}`;
-    const content = readChars(chars, 'both', Source.Lens);
-    const metadata = createPostMetadata(
-        {
-            title,
-            content,
-            marketplace: {
-                name: title,
-                description: content,
-                external_url: SITE_URL,
-            },
-        },
-        createPayloadAttachments(imageResults, videoResult),
-        createPayloadSharingLink(pollResult),
-    );
-
-    const tokenRes = await LensSocialMediaProvider.getAccessToken();
-    const token = tokenRes.unwrap();
-    const arweaveId = await uploadToArweave(metadata, token);
-
-    const commentOn =
-        type === 'reply'
-            ? lensParentPost
-                ? lensParentPost.postId
-                : isThread
-                  ? '$$commentOn$$'
-                  : undefined
-            : undefined;
-
-    return {
-        operationName: resolveLensOperationName(type),
-        variables: {
-            request: {
-                contentURI: `ar://${arweaveId}`,
-                quoteOn: type === 'quote' && lensParentPost ? lensParentPost.postId : undefined,
-                commentOn,
-            },
-        },
-        query: resolveLensQuery(type),
-    };
 }
